+++ CHANGELOG.md
@@ -29,7 +29,14 @@
   - `inherit` is removed
 - These methods of `Sound` has been changed:
   - use `copy` to create a new sound instance
+- These methods of `Sound` has been changed:
+  - `play`, `stop` and `setVolume` method can receive a `duration` parameter
+  - `fade` is removed, use `setVolume` instead
 - In displayable elements, the transform states are separated from the element states
+- These changes are made to `Sound` config
+  - `sync` and `type` are removed
+  - use `preload` to use [Howler.js](https://howlerjs.com/)'s preload feature
+  - use `seek` property to set the initial seek position
 - Scene's config now can't specify the `invertY` and `invertX` properties, use story config `origin` instead
 
 ### Added